be/dev/fix/bookRefactory

See merge request s08-blockchain-nft-sub2/S08P22A308!42

+++ backend/src/main/java/com/ssafy/mongttang/controller/BookController.java
@@ -1,8 +1,9 @@
 package com.ssafy.mongttang.controller;
 
-import com.ssafy.mongttang.dto.ReqSaveBookDto;
-import com.ssafy.mongttang.dto.ReqTemporarySaveBookDto;
+import com.ssafy.mongttang.dto.ReqCreateBookDto;
+import com.ssafy.mongttang.dto.ReqUpdateBookDto;
 import com.ssafy.mongttang.service.BookService;
+import com.ssafy.mongttang.util.TokenUtils;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import lombok.RequiredArgsConstructor;
@@ -13,6 +14,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 import javax.validation.Valid;
 import java.io.IOException;
+import java.security.Principal;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
@@ -32,14 +34,19 @@ public class BookController {
     public ResponseEntity<Map<String,Object>> createBook(@ApiParam(value = "등록된 사진 리스트", required = true, example = "0")
                                                            @RequestPart(value = "imgList", required = false) ArrayList<MultipartFile> imgList,
                                                            @Valid @ApiParam(value = "챌린지 아이디, 동화 제목, 줄거리, 내용, 작가, 완료여부", required = true, example = "0")
-                                                           @RequestPart(value = "BookContent", required = false) ReqSaveBookDto reqSaveBookDto,
+                                                           @RequestPart(value = "BookContent", required = false) ReqCreateBookDto reqCreateBookDto,
                                                            @ApiParam(value = "작가 아이디", required = true, example = "0")
-                                                           @PathVariable int userId){
+                                                           @PathVariable int userId, Principal principal){
         Map<String, Object> resultMap = new HashMap<>();
         HttpStatus status = null;
 
+        if(TokenUtils.compareUserIdAndToken(userId, principal,resultMap)) {
+            status = HttpStatus.BAD_REQUEST;
+            return new ResponseEntity<Map<String, Object>>(resultMap, status);
+        }
+
         try {
-            int bookId = bookService.createBook(userId,reqSaveBookDto,imgList);
+            int bookId = bookService.createBook(userId,reqCreateBookDto,imgList);
             if(bookId > 0){
                 resultMap.put(MESSAGE,SUCCESS);
                 resultMap.put("bookId",bookId);
@@ -61,15 +68,20 @@ public class BookController {
     public ResponseEntity<Map<String,Object>> updateBook(@ApiParam(value = "등록된 사진 리스트", required = true, example = "0")
                                                          @RequestPart(value = "imgList", required = false) ArrayList<MultipartFile> imgList,
                                                          @Valid @ApiParam(value = "챌린지 아이디, 동화 아이디, 동화 제목, 줄거리, 내용, 작가", required = true, example = "0")
-                                                         @RequestPart(value = "BookContent", required = false) ReqTemporarySaveBookDto reqTemporarySaveBookDto,
+                                                         @RequestPart(value = "BookContent", required = false) ReqUpdateBookDto reqUpdateBookDto,
                                                          @ApiParam(value = "작가 아이디", required = true, example = "0")
-                                                         @PathVariable int userId){
+                                                         @PathVariable int userId, Principal principal){
         Map<String, Object> resultMap = new HashMap<>();
         HttpStatus status = null;
 
+        if(TokenUtils.compareUserIdAndToken(userId, principal,resultMap)) {
+            status = HttpStatus.BAD_REQUEST;
+            return new ResponseEntity<Map<String, Object>>(resultMap, status);
+        }
+
         if(imgList != null){
             try {
-                int bookId = bookService.updateBook(userId,reqTemporarySaveBookDto,imgList);
+                int bookId = bookService.updateBook(userId,reqUpdateBookDto,imgList);
                 if(bookId > 0){
                     resultMap.put(MESSAGE,SUCCESS);
                     resultMap.put("bookId",bookId);
@@ -93,9 +105,14 @@ public class BookController {
     @ApiOperation(value = "작가 동화 삭제", notes = "임시저장 한 동화를 삭제한다.", response = Map.class)
     @DeleteMapping("/draw/{userId}")
     public ResponseEntity<Map<String,Object>> deleteBook(@ApiParam(value = "작가 아이디", required = true, example = "0") @PathVariable int userId,
-                                                                 @ApiParam(value = "동화 아이디", required = true, example = "0") @RequestParam int bookId){
+                                                                 @ApiParam(value = "동화 아이디", required = true, example = "0") @RequestParam int bookId, Principal principal){
         Map<String, Object> resultMap = new HashMap<>();
         HttpStatus status = null;
+
+        if(TokenUtils.compareUserIdAndToken(userId, principal,resultMap)) {
+            status = HttpStatus.BAD_REQUEST;
+            return new ResponseEntity<Map<String, Object>>(resultMap, status);
+        }
 
         int isDeleted= bookService.deleteBook(userId,bookId);
